【优化】优化调用方式，真正实现koa2的中间件模式
