+++ package.json
@@ -4,7 +4,7 @@
   "description": "A middleware which can let you send your response easier",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node example.js"
   },
   "repository": {
     "type": "git",
@@ -20,5 +20,9 @@
   "bugs": {
     "url": "https://github.com/detectiveHLH/koa2-response/issues"
   },
-  "homepage": "https://github.com/detectiveHLH/koa2-response#readme"
+  "homepage": "https://github.com/detectiveHLH/koa2-response#readme",
+  "dependencies": {
+    "koa": "^2.5.2",
+    "koa-router": "^7.4.0"
+  }
 }
